ical2jsonld: convert ical events to JSON-LD structured data, for crawlers.
Google has the ability to show local events in a special pane, more
convenient (and maybe more attractive) to users: https://kalicube.com/wp-content/uploads/2022/03/Event-Box-Google-SERP-2.png

Foulab events have been seen to render this way some times, and other
times they don't.

Google specifies a structured data format for events:
https://developers.google.com/search/docs/appearance/structured-data/event
(which is part of the broader schema.org standard)

This commit makes the website dynamically convert the existing
foulab.ics iCalendar file into JSON-LD structured data, in the hopes
that this will make our events show up consistently in Google search
results. (For example, imagine someone searching for 'montreal
electronics events' and seeing an event at Foulab.)

The generated JSON-LD structured data is invisible to ordinary visitors.
In fact, to save their network/CPU resources, the conversion code runs
only when it is the Google crawler visiting. (The question of whether
Google would be OK with this was discussed here:
https://github.com/FOULAB/foulab-dot-org-hugo/commit/648de2a368bb3abb23119c6ca63faa8c6556285d#r142974964)

When publishing new events for Foulab, the process is the same as before:
modify the foulab.ics file and the markdown files in content/. (because the
JSON-LD is generated from the foulab.ics file)